+++ ext/tiny_tds/client.c
@@ -4,7 +4,6 @@
 
 VALUE cTinyTdsClient;
 extern VALUE mTinyTds, cTinyTdsError;
-static ID sym_username, sym_password, sym_dataserver, sym_database, sym_appname, sym_tds_version, sym_login_timeout, sym_timeout, sym_encoding, sym_azure, sym_contained, sym_use_utf16, sym_message_handler;
 static ID intern_source_eql, intern_severity_eql, intern_db_error_number_eql, intern_os_error_number_eql;
 static ID intern_new, intern_dup, intern_transpose_iconv_encoding, intern_local_offset, intern_gsub, intern_call;
 VALUE opt_escape_regex, opt_escape_dblquote;
@@ -470,7 +469,7 @@ static VALUE allocate(VALUE klass)
 
 // TinyTds::Client (public)
 
-static VALUE rb_tinytds_tds_version(VALUE self)
+static VALUE rb_tinytds_server_version(VALUE self)
 {
   GET_CLIENT_WRAPPER(self);
   return INT2FIX(dbtds(cwrap->client));
@@ -925,25 +924,25 @@ static VALUE rb_tinytds_identity_sql(VALUE self)
 
 // TinyTds::Client (protected)
 
-static VALUE rb_tinytds_connect(VALUE self, VALUE opts)
+static VALUE rb_tinytds_connect(VALUE self)
 {
   /* Parsing options hash to local vars. */
-  VALUE user, pass, dataserver, database, app, version, ltimeout, timeout, charset, azure, contained, use_utf16;
+  VALUE username, password, dataserver, database, app_name, tds_version, login_timeout, timeout, charset, azure, contained, use_utf16;
   GET_CLIENT_WRAPPER(self);
 
-  user = rb_hash_aref(opts, sym_username);
-  pass = rb_hash_aref(opts, sym_password);
-  dataserver = rb_hash_aref(opts, sym_dataserver);
-  database = rb_hash_aref(opts, sym_database);
-  app = rb_hash_aref(opts, sym_appname);
-  version = rb_hash_aref(opts, sym_tds_version);
-  ltimeout = rb_hash_aref(opts, sym_login_timeout);
-  timeout = rb_hash_aref(opts, sym_timeout);
-  charset = rb_hash_aref(opts, sym_encoding);
-  azure = rb_hash_aref(opts, sym_azure);
-  contained = rb_hash_aref(opts, sym_contained);
-  use_utf16 = rb_hash_aref(opts, sym_use_utf16);
-  cwrap->userdata->message_handler = rb_hash_aref(opts, sym_message_handler);
+  app_name = rb_iv_get(self, "@app_name");
+  azure = rb_iv_get(self, "@azure");
+  contained = rb_iv_get(self, "@contained");
+  database = rb_iv_get(self, "@database");
+  dataserver = rb_iv_get(self, "@dataserver");
+  charset = rb_iv_get(self, "@encoding");
+  login_timeout = rb_iv_get(self, "@login_timeout");
+  password = rb_iv_get(self, "@password");
+  tds_version = rb_iv_get(self, "@tds_version");
+  timeout = rb_iv_get(self, "@timeout");
+  username = rb_iv_get(self, "@username");
+  use_utf16 = rb_iv_get(self, "@use_utf16");
+  cwrap->userdata->message_handler = rb_iv_get(self, "@message_handler");
 
   /* Dealing with options. */
   if (dbinit() == FAIL) {
@@ -955,24 +954,24 @@ static VALUE rb_tinytds_connect(VALUE self, VALUE opts)
   dbmsghandle(tinytds_msg_handler);
   cwrap->login = dblogin();
 
-  if (!NIL_P(version)) {
-    dbsetlversion(cwrap->login, NUM2INT(version));
+  if (!NIL_P(tds_version)) {
+    dbsetlversion(cwrap->login, NUM2INT(tds_version));
   }
 
-  if (!NIL_P(user)) {
-    dbsetluser(cwrap->login, StringValueCStr(user));
+  if (!NIL_P(username)) {
+    dbsetluser(cwrap->login, StringValueCStr(username));
   }
 
-  if (!NIL_P(pass)) {
-    dbsetlpwd(cwrap->login, StringValueCStr(pass));
+  if (!NIL_P(password)) {
+    dbsetlpwd(cwrap->login, StringValueCStr(password));
   }
 
-  if (!NIL_P(app)) {
-    dbsetlapp(cwrap->login, StringValueCStr(app));
+  if (!NIL_P(app_name)) {
+    dbsetlapp(cwrap->login, StringValueCStr(app_name));
   }
 
-  if (!NIL_P(ltimeout)) {
-    dbsetlogintime(NUM2INT(ltimeout));
+  if (!NIL_P(login_timeout)) {
+    dbsetlogintime(NUM2INT(login_timeout));
   }
 
   if (!NIL_P(charset)) {
@@ -981,19 +980,7 @@ static VALUE rb_tinytds_connect(VALUE self, VALUE opts)
 
   if (!NIL_P(database)) {
     if (azure == Qtrue || contained == Qtrue) {
-      #ifdef DBSETLDBNAME
       DBSETLDBNAME(cwrap->login, StringValueCStr(database));
-      #else
-
-      if (azure == Qtrue) {
-        rb_warn("TinyTds: :azure option is not supported in this version of FreeTDS.\n");
-      }
-
-      if (contained == Qtrue) {
-        rb_warn("TinyTds: :contained option is not supported in this version of FreeTDS.\n");
-      }
-
-      #endif
     }
   }
 
@@ -1017,8 +1004,8 @@ static VALUE rb_tinytds_connect(VALUE self, VALUE opts)
     cwrap->closed = 0;
     cwrap->charset = charset;
 
-    if (!NIL_P(version)) {
-      dbsetversion(NUM2INT(version));
+    if (!NIL_P(tds_version)) {
+      dbsetversion(NUM2INT(tds_version));
     }
 
     if (!NIL_P(timeout)) {
@@ -1053,7 +1040,7 @@ void init_tinytds_client()
   cTinyTdsClient = rb_define_class_under(mTinyTds, "Client", rb_cObject);
   rb_define_alloc_func(cTinyTdsClient, allocate);
   /* Define TinyTds::Client Public Methods */
-  rb_define_method(cTinyTdsClient, "tds_version", rb_tinytds_tds_version, 0);
+  rb_define_method(cTinyTdsClient, "server_version", rb_tinytds_server_version, 0);
   rb_define_method(cTinyTdsClient, "close", rb_tinytds_close, 0);
   rb_define_method(cTinyTdsClient, "closed?", rb_tinytds_closed, 0);
   rb_define_method(cTinyTdsClient, "canceled?", rb_tinytds_canceled, 0);
@@ -1068,21 +1055,7 @@ void init_tinytds_client()
   rb_define_method(cTinyTdsClient, "return_code", rb_tinytds_return_code, 0);
   rb_define_method(cTinyTdsClient, "identity_sql", rb_tinytds_identity_sql, 0);
   /* Define TinyTds::Client Protected Methods */
-  rb_define_protected_method(cTinyTdsClient, "connect", rb_tinytds_connect, 1);
-  /* Symbols For Connect */
-  sym_username = ID2SYM(rb_intern("username"));
-  sym_password = ID2SYM(rb_intern("password"));
-  sym_dataserver = ID2SYM(rb_intern("dataserver"));
-  sym_database = ID2SYM(rb_intern("database"));
-  sym_appname = ID2SYM(rb_intern("appname"));
-  sym_tds_version = ID2SYM(rb_intern("tds_version"));
-  sym_login_timeout = ID2SYM(rb_intern("login_timeout"));
-  sym_timeout = ID2SYM(rb_intern("timeout"));
-  sym_encoding = ID2SYM(rb_intern("encoding"));
-  sym_azure = ID2SYM(rb_intern("azure"));
-  sym_contained = ID2SYM(rb_intern("contained"));
-  sym_use_utf16 = ID2SYM(rb_intern("use_utf16"));
-  sym_message_handler = ID2SYM(rb_intern("message_handler"));
+  rb_define_protected_method(cTinyTdsClient, "connect", rb_tinytds_connect, 0);
   /* Intern TinyTds::Error Accessors */
   intern_source_eql = rb_intern("source=");
   intern_severity_eql = rb_intern("severity=");