+++ test/client_test.rb
@@ -26,14 +26,9 @@ class ClientTest < TinyTds::TestCase
       end
     end
 
-    it "has getters for the tds version information (brittle since conf takes precedence)" do
-      if @client.tds_73?
-        assert_equal 11, @client.tds_version
-        assert_equal "DBTDS_7_3 - Microsoft SQL Server 2008", @client.tds_version_info
-      else
-        assert_equal 9, @client.tds_version
-        assert_equal "DBTDS_7_1/DBTDS_8_0 - Microsoft SQL Server 2000", @client.tds_version_info
-      end
+    it "has getters for the server version information (brittle since conf takes precedence)" do
+      assert_equal 11, @client.server_version
+      assert_equal "DBTDS_7_3 - Microsoft SQL Server 2008", @client.server_version_info
     end
 
     it "uses UTF-8 client charset/encoding by default" do
@@ -82,8 +77,7 @@ class ClientTest < TinyTds::TestCase
     end
 
     it "raises TinyTds exception with undefined :dataserver" do
-      options = connection_options login_timeout: 1, dataserver: "DOESNOTEXIST"
-      action = lambda { new_connection(options) }
+      action = lambda { new_connection(login_timeout: 1, dataserver: "DOESNOTEXIST") }
       assert_raise_tinytds_error(action) do |e|
         # Not sure why tese are different.
         if ruby_darwin?
@@ -193,7 +187,7 @@ class ClientTest < TinyTds::TestCase
     it "raises TinyTds exception with wrong :username" do
       skip if ENV["CI"] && sqlserver_azure? # Some issue with db_error_number.
       options = connection_options username: "willnotwork"
-      action = lambda { new_connection(options) }
+      action = lambda { new_connection(**options) }
       assert_raise_tinytds_error(action) do |e|
         assert_equal 18456, e.db_error_number
         assert_equal 14, e.severity